Added AccountWrapper component
- This component will set the current account and allows for
  subcomponents to simply query the current account rather than dealing
  with the endpoint itself.
