+++ src/features/selectors.ts
@@ -1,0 +1,9 @@
+import { selectors as account } from "./account";
+import { selectors as api } from "./api";
+import { selectors as holdings } from "./holdings";
+
+export const selectors = {
+  account,
+  api,
+  holdings,
+};
